+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFWV1XHQCNBPJ72V4VDBBNW3"
+        "01KFX4XN6FKCTGSFF0T591FRJW"
     )
     r"""The unique identifier of the data source"""