+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFWV1XHT9TPX6VNT3K2HJN5X"
+        "01KFX4XN6HPN21FGW3KJ7GMCEZ"
     )
     r"""The unique identifier of the data source"""