+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-26T09:44:29.020Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-26T12:36:54.881Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")