+++ packages/orq-rc/RELEASES.md
@@ -5049,3 +5049,13 @@ Based on:
 - [python v4.3.0-rc.2] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.2 - packages/orq-rc
+
+## 2026-01-26 12:39:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.697.5 (2.799.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.3] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.3] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.3 - packages/orq-rc